+++ CHANGELOG.md
@@ -16,6 +16,8 @@ x.x.x Release notes (yyyy-MM-dd)
   names (e.g. `@objc(Foo) class SwiftFoo: Object {}`).
 * Fix `-[RLMMigration enumerateObjects:block:]` returning incorrect `oldObject`
   objects when enumerating a class name after previously deleting a `newObject`.
+* Fix an issue where `Realm.asyncOpen(...)` would fail to work when opening a
+  synchronized Realm for which the user only had read permissions.
 
 2.6.2 Release notes (2017-04-21)
 =============================================================